renamed orbitviewitem
