+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/OrbitView/OrbitViewXaml.bind
@@ -7,7 +7,7 @@
                     AnchorColor="@[AnchorColor:Brush:Gray]"
                     OrbitColor="@[OrbitColor:Brush:Gray]">
   <controls:OrbitView.ItemTemplate>
-    <DataTemplate x:DataType="controls:OrbitViewItem">
+    <DataTemplate x:DataType="controls:OrbitViewDataItem">
       <controls:DropShadowPanel Color="Black" BlurRadius="20" VerticalContentAlignment="Stretch" HorizontalContentAlignment="Stretch">
         <Ellipse >
           <Ellipse.Fill>
@@ -18,12 +18,12 @@
     </DataTemplate>
   </controls:OrbitView.ItemTemplate>
   <controls:OrbitView.ItemsSource>
-    <controls:OrbitViewItemCollection>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/shen.png" Distance="0.1" Label="Shen" Diameter="0.2"></controls:OrbitViewItem>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/david.png" Distance="0.2" Label="David" Diameter="0.5"></controls:OrbitViewItem>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/petri.png" Distance="0.4" Label="Petri" Diameter="0.6"></controls:OrbitViewItem>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/vlad.png" Distance="0.8" Label="Vlad" Diameter="0.8"></controls:OrbitViewItem>
-    </controls:OrbitViewItemCollection>
+    <controls:OrbitViewDataItemCollection>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/shen.png" Distance="0.1" Label="Shen" Diameter="0.2"></controls:OrbitViewDataItem>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/david.png" Distance="0.2" Label="David" Diameter="0.5"></controls:OrbitViewDataItem>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/petri.png" Distance="0.4" Label="Petri" Diameter="0.6"></controls:OrbitViewDataItem>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/vlad.png" Distance="0.8" Label="Vlad" Diameter="0.8"></controls:OrbitViewDataItem>
+    </controls:OrbitViewDataItemCollection>
   </controls:OrbitView.ItemsSource>
   <controls:OrbitView.CenterContent>
     <Grid>
