+++ Microsoft.Toolkit.Uwp.UI.Controls/Microsoft.Toolkit.Uwp.UI.Controls.csproj
@@ -111,16 +111,16 @@
     <Compile Include="MasterDetailsView\MasterDetailsView.Properties.cs" />
     <Compile Include="BladeView\BladeItem.Events.cs" />
     <Compile Include="MasterDetailsView\MasterDetailsViewState.cs" />
-    <Compile Include="OrbitView\OrbitViewItemControl.cs" />
+    <Compile Include="OrbitView\OrbitViewItem.cs" />
     <Compile Include="ScrollHeader\ScrollHeaderMode.cs" />
     <Compile Include="RadialProgressBar\RadialProgressBar.cs" />
     <Compile Include="SlidableListItem\SwipeStatus.cs" />
     <Compile Include="SlidableListItem\SwipeStatusChangedEventArgs.cs" />
     <Compile Include="OrbitView\OrbitView.cs" />
     <Compile Include="OrbitView\OrbitViewElementProperties.cs" />
-    <Compile Include="OrbitView\OrbitViewItem.cs" />
+    <Compile Include="OrbitView\OrbitViewDataItem.cs" />
     <Compile Include="OrbitView\OrbitViewItemClickedEventArgs.cs" />
-    <Compile Include="OrbitView\OrbitViewItemCollection.cs" />
+    <Compile Include="OrbitView\OrbitViewDataItemCollection.cs" />
     <Compile Include="OrbitView\OrbitViewPanel.cs" />
     <Compile Include="OrbitView\OrbitViewPanelItemArrangedArgs.cs" />
     <Compile Include="OrbitView\OrbitViewPanelItemsArrangedArgs.cs" />
