+++ docs/controls/OrbitView.md
@@ -16,7 +16,7 @@ The `OrbitView` control provides a new control, inherited from the `ItemsControl
                     AnchorColor="Gray"
                     OrbitColor="Gray">
   <controls:OrbitView.ItemTemplate>
-    <DataTemplate x:DataType="controls:OrbitViewItem">
+    <DataTemplate x:DataType="controls:OrbitViewDataItem">
       <controls:DropShadowPanel Color="Black" BlurRadius="20" VerticalContentAlignment="Stretch" HorizontalContentAlignment="Stretch">
         <Ellipse >
           <Ellipse.Fill>
@@ -27,12 +27,12 @@ The `OrbitView` control provides a new control, inherited from the `ItemsControl
     </DataTemplate>
   </controls:OrbitView.ItemTemplate>
   <controls:OrbitView.ItemsSource>
-    <controls:OrbitViewItemCollection>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/shen.png" Distance="0.1" Label="Shen" Diameter="0.2"></controls:OrbitViewItem>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/david.png" Distance="0.2" Label="David" Diameter="0.5"></controls:OrbitViewItem>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/petri.png" Distance="0.4" Label="Petri" Diameter="0.6"></controls:OrbitViewItem>
-      <controls:OrbitViewItem Image="ms-appx:///Assets/People/vlad.png" Distance="0.8" Label="Vlad" Diameter="0.8"></controls:OrbitViewItem>
-    </controls:OrbitViewItemCollection>
+    <controls:OrbitViewDataItemCollection>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/shen.png" Distance="0.1" Label="Shen" Diameter="0.2"></controls:OrbitViewDataItem>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/david.png" Distance="0.2" Label="David" Diameter="0.5"></controls:OrbitViewDataItem>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/petri.png" Distance="0.4" Label="Petri" Diameter="0.6"></controls:OrbitViewDataItem>
+      <controls:OrbitViewDataItem Image="ms-appx:///Assets/People/vlad.png" Distance="0.8" Label="Vlad" Diameter="0.8"></controls:OrbitViewDataItem>
+    </controls:OrbitViewDataItemCollection>
   </controls:OrbitView.ItemsSource>
   <controls:OrbitView.CenterContent>
     <Grid>
@@ -79,12 +79,12 @@ When Anchors are enabled, this property specifies the anchor line thickness
 ### MinItemSize
 MinItemSize specifies the minimum size for items
 
-> Note: for this property to work, the data context of each item has to be a `OrbitViewItem` and Diameter property set between 0 and 1 
+> Note: for this property to work, the data context of each item has to be a `OrbitViewDataItem` and Diameter property set between 0 and 1 
 
 ### MaxItemSize
 MinItemSize specifies the maximum size for items
 
-> Note: for this property to work, the data context of each item has to be a `OrbitViewItem` and Diameter property set between 0 and 1 
+> Note: for this property to work, the data context of each item has to be a `OrbitViewDataItem` and Diameter property set between 0 and 1 
 
 ## OrbitView Events
 
@@ -93,8 +93,8 @@ Event is raised when item is invoked with mouse or keyboard/controller
 
 > Note: `IsItemClickedEnabled` should be true for this event to work
 
-## OrbitViewItem
-OrbitViewItem is a helper class used for specifying size and distance of each item of the `OrbitView`. To work properly, the `ItemSource` of the `OrbitView` should be set to en *IEnumerable< OrbitViewItem >*. Objects extending OrbitViewItem will also work. Alternatively, OrbitViewItem has the `Item` object property that can be used to store additional objects and properties.
+## OrbitViewDataItem
+OrbitViewDataItem is a helper class used for specifying size and distance of each item of the `OrbitView`. To work properly, the `ItemSource` of the `OrbitView` should be set to en *IEnumerable< OrbitViewDataItem >*. Objects extending OrbitViewDataItem will also work. Alternatively, OrbitViewDataItem has the `Item` object property that can be used to store additional objects and properties.
 
 ### Diameter
 A double value between 0 and 1, 0 indicating `MinItemSize` and 1 indicating `MaxItemSize`
